# testovoe/api-classes/HistogramApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package o.e.api;

import java.util.Map;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-09T20:58:53.400428235+03:00[Europe/Moscow]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "hotels", description = "the hotels API")
public interface HistogramApi {

    String PATH_GET_HISTOGRAM = "/histogram/{param}";
    /**
     * GET /histogram/{param} : Get count of hotels grouped by parameter value
     *
     * @param param  (required)
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "getHistogram",
        summary = "Get count of hotels grouped by parameter value",
        tags = { "hotels" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Map.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = HistogramApi.PATH_GET_HISTOGRAM,
        produces = { "application/json" }
    )
    ResponseEntity<Map<String, Integer>> getHistogram(
        @NotNull @Parameter(name = "param", description = "", required = true, in = ParameterIn.PATH) @PathVariable("param") String param
    );

}