# testovoe/api-classes/SearchApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package o.e.api;

import o.e.api.model.HotelBrief;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-09T20:58:53.400428235+03:00[Europe/Moscow]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "hotels", description = "the hotels API")
public interface SearchApi {

    String PATH_SEARCH_HOTELS = "/search";
    /**
     * GET /search : Search hotels by parameters
     *
     * @param name  (optional)
     * @param brand  (optional)
     * @param city  (optional)
     * @param country  (optional)
     * @param amenities  (optional)
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "searchHotels",
        summary = "Search hotels by parameters",
        tags = { "hotels" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = HotelBrief.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SearchApi.PATH_SEARCH_HOTELS,
        produces = { "application/json" }
    )
    ResponseEntity<List<HotelBrief>> searchHotels(
        @Parameter(name = "name", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "name", required = false) @Nullable String name,
        @Parameter(name = "brand", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "brand", required = false) @Nullable String brand,
        @Parameter(name = "city", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "city", required = false) @Nullable String city,
        @Parameter(name = "country", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "country", required = false) @Nullable String country,
        @Parameter(name = "amenities", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "amenities", required = false) @Nullable String amenities
    );

}